refactor: clear code health degradations on branch changes (CS-11648)
Extract helpers to reduce complexity in codescene file listeners, git ignore
batching, change lister scan logic, and process execution; split oversized
test suites into focused files.
